+++ src/Data/MapItemTrait.php
@@ -123,5 +123,4 @@ trait MapItemTrait
         }
         return null;
     }
-
 }