+++ src/Data/Repository/RestApi/RestApi.php
@@ -137,5 +137,4 @@ class RestApi extends ContentRepository  //implements RepositoryInterface
         $data = $this->getProvider()->decode($response);
         return $mapper->map($data, $contentPropertyPath);
     }
-
 }